+++ Topos.Kafka/Kafka/KafkaProducerImplementation.cs
@@ -16,32 +16,54 @@ namespace Topos.Kafka
         static readonly Headers EmptyHeaders = new Headers();
 
         readonly IProducer<string, byte[]> _producer;
+        readonly ProducerConfig _config;
         readonly ILogger _logger;
-        readonly string _address;
 
         bool _disposed;
 
-        public KafkaProducerImplementation(ILoggerFactory loggerFactory, string address, int sendTimeoutSeconds = 30)
+        public KafkaProducerImplementation(ILoggerFactory loggerFactory, string address, int sendTimeoutSeconds = 30, Func<ProducerConfig, ProducerConfig> configurationCustomizer = null)
         {
-            _address = address;
             _logger = loggerFactory.GetLogger(typeof(KafkaProducerImplementation));
 
             var config = new ProducerConfig
             {
                 BootstrapServers = address,
                 MessageTimeoutMs = sendTimeoutSeconds * 1000,
-                LogConnectionClose = false
+                LogConnectionClose = false,
             };
+
+            if (configurationCustomizer != null)
+            {
+                config = configurationCustomizer(config);
+            }
 
             _producer = new ProducerBuilder<string, byte[]>(config)
                 .SetLogHandler((producer, message) => LogHandler(_logger, producer, message))
                 .SetErrorHandler((producer, message) => ErrorHandler(_logger, producer, message))
                 .Build();
 
-            _logger.Info("Kafka producer initialized with {address}", address);
+            _logger.Info("Kafka producer initialized with {address}", config.BootstrapServers);
+
+            _config = config;
         }
 
-        public IAdminClient GetAdminClient() => new AdminClientBuilder(new[] { new KeyValuePair<string, string>("bootstrap.servers", _address) }).Build();
+        public IAdminClient GetAdminClient()
+        {
+            // sasl.jaas.config=org.apache.kafka.common.security.plain.PlainLoginModule required username="(username)" password="(password)";
+
+            var configuration = new List<KeyValuePair<string, string>>
+            {
+                new KeyValuePair<string, string>("bootstrap.servers", _config.BootstrapServers),
+            };
+
+            if (!string.IsNullOrWhiteSpace(_config.SaslUsername))
+            {
+                configuration.Add(new KeyValuePair<string, string>("sasl.username", _config.SaslUsername));
+                configuration.Add(new KeyValuePair<string, string>("sasl.password", _config.SaslPassword));
+            }
+
+            return new AdminClientBuilder(configuration).Build();
+        }
 
         public async Task Send(string topic, string partitionKey, TransportMessage transportMessage)
         {