+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
-## 0.0.27
-* Confluent Kafka 1.1.0* Kafkaesque-based broker implementation (for the file system)
+## 0.0.28
+* Confluent Kafka 1.2.0
+* Kafkaesque-based broker implementation (for the file system)
 * Ability to destructure objects, Serilog style
+* Can use Azure Event Hubs connection string w. Kafka transport (automatically parses it & sets the correct parameters)